requirements.txt
python-docx==1.1.2
fastapi==0.115.4
uvicorn==0.32.0
python-multipart==0.0.17